requirements.txt
azure-functions
firebase-admin
google-cloud-firestore
google-api-core==1.17.0
google-cloud-core==1.3.0